fix: Address clippy warnings for CI compliance
- Add #[allow(dead_code)] to unused stop_container fn
- Replace redundant closures with function references
- Rename JobPhase::from_str to parse to avoid trait confusion
- Add #[allow(clippy::large_enum_variant)] for proto module
- Fix wildcard pattern in CAS client match
- Remove unnecessary u16 cast in docker_cli
- Use &Path instead of &PathBuf in macos.rs

+++ src/runtime/docker_cli.rs
@@ -379,7 +379,7 @@ impl Runtime for DockerCliRuntime {
             pid: Some(pid),
             bundle_path: Some(PathBuf::from("/")),
             log_path: Some(log_path),
-            port: Some(host_port as u16),
+            port: Some(host_port),
         })
     }
